requirements.txt
duckdb>=1.0
pyarrow>=15.0
langchain>=0.3
langchain-community>=0.3
langchain-ollama>=0.2
chainlit>=1.1
ollama>=0.3
pyyaml>=6.0
plotly>=5.18
pandas>=2.0
openpyxl>=3.1
reportlab>=4.0
